+++ include/BDSBeamPipeFactoryBase.hh
@@ -127,8 +127,9 @@ protected:
   G4double         lengthSafetyLarge;
   
   G4bool           checkOverlaps;
-  G4double         maxStepFactor;
   G4double         nSegmentsPerCircle; ///< For visualisation improvement
+  G4bool           sensitiveBeamPipe;  ///< Whether the beam pipe will record energy deposition.
+  G4bool           sensitiveVacuum;    ///< Wehther the vacuum will record any energy deposition.
   G4VSolid*        vacuumSolid;
   G4VSolid*        beamPipeSolid;
   G4VSolid*        containerSolid;