+++ include/BDSOutputROOTEventSampler.hh
@@ -65,6 +65,7 @@ public:
 
   /// @{ These are not filled by default.
   std::vector<int>     charge;
+  std::vector<U>       kineticEnergy;
   std::vector<U>       mass;
   std::vector<U>       rigidity;
   std::vector<bool>    isIon;
@@ -74,6 +75,7 @@ public:
 
   /// @{ Function to calculate on the fly the parameters.
   std::vector<int>     getCharge();
+  std::vector<U>       getKineticEnergy();
   std::vector<U>       getMass();
   std::vector<U>       getRigidity();
   std::vector<bool>    getIsIon();
@@ -98,17 +100,20 @@ public:
   inline void FillMass()     {mass     = getMass();}
   inline void FillRigidity() {rigidity = getRigidity();}
   inline void FillIon()      {isIon = getIsIon(); ionA = getIonA(); ionZ = getIonZ();}
+  inline void FillKineticEnergy() {kineticEnergy = getKineticEnergy();}
   /// @}
   
   void FillCMR();  ///< Calculate and fill charge, mass and rigidity.
+  void FillCMRK(); ///< Calculate and fill charge, mass, rigidity, and kinetic energy.
   void FillCMRI(); ///< Calculate and fill charge, mass, rigidity and ion properties.
-  
+  void FillCMRIK(); ///< Calculate and fill charge, mass, rigidity, kinetic energy, and ion properties.
+
   void SetBranchAddress(TTree *);
   void Flush();  ///< Clean Sampler
 
   static BDSOutputROOTGeant4Data* particleTable;
 
-  ClassDef(BDSOutputROOTEventSampler,2);
+  ClassDef(BDSOutputROOTEventSampler,3);
 };
 
 #endif