+++ parser/options.cc
@@ -219,7 +219,6 @@ void Options::PublishMembers()
   publish("tunnelAper1",         &Options::tunnelAper1);
   publish("tunnelAper2",         &Options::tunnelAper2);
   publish("tunnelRadius",        &Options::tunnelAper1); // for backwards compatability
-  publish("tunnelSensitive",     &Options::tunnelSensitive);
   publish("tunnelVisible",       &Options::tunnelVisible);
   publish("showTunnel",          &Options::tunnelVisible); // for backwards compatability
   publish("tunnelOffsetX",       &Options::tunnelOffsetX);
@@ -233,7 +232,7 @@ void Options::PublishMembers()
   publish("blmRad",    &Options::blmRad);
   publish("blmLength", &Options::blmLength);
 
-    // physics processes
+  // physics processes
   publish("turnOnOpticalAbsorption",     &Options::turnOnOpticalAbsorption);
   publish("turnOnMieScattering",         &Options::turnOnMieScattering);
   publish("turnOnRayleighScattering",    &Options::turnOnRayleighScattering);
@@ -244,8 +243,12 @@ void Options::PublishMembers()
   publish("maximumTracksPerEvent",       &Options::maximumTracksPerEvent);
   publish("minimumKineticEnergy",        &Options::minimumKineticEnergy);
   publish("minimumRange",                &Options::minimumRange);
-  publish("sensitiveBeamlineComponents", &Options::sensitiveBeamlineComponents);
+  publish("sensitiveOuter",              &Options::sensitiveOuter);
+  publish("sensitiveBeamlineComponents", &Options::sensitiveOuter); // backwards compatibility
   publish("sensitiveBeamPipe",           &Options::sensitiveBeamPipe);
+  publish("sensitiveTunnel",             &Options::sensitiveTunnel);
+  publish("tunnelSensitive",             &Options::sensitiveTunnel);// backwards compatibility
+  publish("sensitiveVacuum",             &Options::sensitiveVacuum);
   publish("sensitiveBLMs",               &Options::sensitiveBLMs);
   publish("prodCutPhotons",              &Options::prodCutPhotons);
   publish("prodCutElectrons",            &Options::prodCutElectrons);
@@ -291,6 +294,7 @@ void Options::PublishMembers()
   publish("trajCutGTZ",                     &Options::trajCutGTZ);
   publish("trajCutLTR",                     &Options::trajCutLTR);
   publish("trajNoTransportation",           &Options::trajNoTransportation);
+  publish("storeElossTurn",                 &Options::storeElossTurn);
   publish("storeElossLinks",                &Options::storeElossLinks);
   publish("storeElossLocal",                &Options::storeElossLocal);
   publish("storeElossGlobal",               &Options::storeElossGlobal);
@@ -306,6 +310,7 @@ void Options::PublishMembers()
   publish("storeTrajectorySamplerID"      , &Options::storeTrajectorySamplerID);
   publish("storeTrajectoryELossSRange"    , &Options::storeTrajectoryELossSRange);
   publish("storeSamplerCharge",             &Options::storeSamplerCharge);
+  publish("storeSamplerKineticEnergy",      &Options::storeSamplerKineticEnergy);
   publish("storeSamplerMass",               &Options::storeSamplerMass);
   publish("storeSamplerRigidity",           &Options::storeSamplerRigidity);
   publish("storeSamplerIon",                &Options::storeSamplerIon);