+++ src/BDSEventAction.cc
@@ -32,6 +32,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "BDSTerminatorSD.hh"
 #include "BDSTrajectory.hh"
 #include "BDSTrajectoryPrimary.hh"
+#include "BDSUtilities.hh"
 
 #include "globals.hh"                  // geant4 types / globals
 #include "G4Event.hh"
@@ -169,15 +170,14 @@ void BDSEventAction::EndOfEventAction(const G4Event* evt)
   // Get the current time
   stopTime = time(nullptr);
   eventInfo->SetStopTime(stopTime);
-  
-  // Get the curent memory usage
-  struct rusage r_usage;
-  getrusage(RUSAGE_SELF,&r_usage);
-  eventInfo->SetMemoryUsage(r_usage.ru_maxrss/(1048*1048));
 
+  // Timing information
   milliseconds ms = duration_cast<milliseconds>(system_clock::now().time_since_epoch());
   stops = (G4double)ms.count()/1000.0;
   eventInfo->SetDuration(G4float(stops - starts));
+
+  G4double memoryUsedMb = BDS::GetMemoryUsage();
+  eventInfo->SetMemoryUsage(memoryUsedMb);
 
   // Get the hits collection of this event - all hits from different SDs.
   G4HCofThisEvent* HCE = evt->GetHCofThisEvent();